deleting old designspace builder
