+++ Tools/build-designspace.py
@@ -4,16 +4,6 @@ reload(xTools4.modules.sys)
 import xTools4.modules.measurements
 reload(xTools4.modules.measurements)
 
-'''
-this script needs to ne cleaned-up and/or rewritten:
-
-- avar1 designspace is no longer in use
-- what about fences? current implementation is not correct
-- avar2.v2 is probably no longer needed too
-- variable font generation is now done in Terminal with build.sh
-
-'''
-
 import os, glob, shutil, json, time, datetime
 import subprocess
 from xml.etree.ElementTree import parse
@@ -30,35 +20,154 @@ from xTools4.modules.linkPoints2 import readMeasurements
 from xTools4.modules.sys import timer
 
 
-SUBFAMILY = ['Roman', 'Italic'][1]
+def makeParentAxis(parentName, parametricAxes, defaultName):
+    r'''
+    Calculate a parent axis to control several parametric axes, 
+    with mappings to limit the range of each child axis.
 
-ASCII  = 'space exclam quotedbl numbersign dollar percent ampersand quotesingle parenleft parenright asterisk plus comma hyphen period slash zero one two three four five six seven eight nine colon semicolon less equal greater question at A B C D E F G H I J K L M N O P Q R S T U V W X Y Z bracketleft backslash bracketright asciicircum underscore grave a b c d e f g h i j k l m n o p q r s t u v w x y z braceleft bar braceright asciitilde'
-LATIN1 = ASCII + ' exclamdown cent sterling currency yen brokenbar section dieresis copyright ordfeminine guillemotleft logicalnot registered macron degree plusminus twosuperior threesuperior acute uni00B5 micro paragraph periodcentered cedilla onesuperior ordmasculine guillemotright onequarter onehalf threequarters questiondown Agrave Aacute Acircumflex Atilde Adieresis Aring AE Ccedilla Egrave Eacute Ecircumflex Edieresis Igrave Iacute Icircumflex Idieresis Eth Ntilde Ograve Oacute Ocircumflex Otilde Odieresis multiply Oslash Ugrave Uacute Ucircumflex Udieresis Yacute Thorn germandbls agrave aacute acircumflex atilde adieresis aring ae ccedilla egrave eacute ecircumflex edieresis igrave iacute icircumflex idieresis eth ntilde ograve oacute ocircumflex otilde odieresis divide oslash ugrave uacute ucircumflex udieresis yacute thorn ydieresis idotless Lslash lslash OE oe Scaron scaron Ydieresis Zcaron zcaron florin circumflex caron breve dotaccent ring ogonek tilde hungarumlaut endash emdash quoteleft quoteright quotesinglbase quotedblleft quotedblright quotedblbase dagger daggerdbl bullet ellipsis perthousand guilsinglleft guilsinglright fraction Euro trademark minus fi fl'
+    ::
+        parentName  = 'XTRA'
+        parametricAxes = {
+            'XTUC' : dict(minimum=72, maximum=668, default=400),
+            'XTUR' : dict(minimum=60, maximum=902, default=561),
+            'XTUD' : dict(minimum=76, maximum=686, default=410),
+            'XTLC' : dict(minimum=42, maximum=500, default=243),
+            'XTLR' : dict(minimum=46, maximum=625, default=337),
+            'XTLD' : dict(minimum=84, maximum=501, default=248),
+            'XTFI' : dict(minimum=40, maximum=604, default=329),
+        }
+        defaultName = 'XTUC'
+
+        parentAxis, mappings = makeParentAxis(parentName, parametricAxes, defaultName)
+        
+        print('parent parametric axis:')
+        print(parentAxis)
+        print()
+        print('parent mappings to child parameters:')
+        for parentValue, mapping in sorted(mappings.items()):
+            print(f'\t{ parentValue } { mapping }')
+
+    '''
+    # THIS IS THE WRONG PLACE FOR THIS KIND OF DATA!
+    # MOVE TO DESIGNSPACE BUILDER? MEASUREMENTS FORMAT?
+    matchRangeAxes = {
+        'XQUC' : 'XTUR',
+        'XQLC' : 'XTLR',
+        'XQFI' : 'XTFI',
+    }
+
+    defaultValue = parametricAxes[defaultName]['default']
+    minValues = []
+    maxValues = []
+    for axisName, axis in parametricAxes.items():
+        # SKIP MATCHED RANGE AXES
+        if axisName in matchRangeAxes:
+            continue
+        axisShift = defaultValue - axis['default'] 
+        minValue  = axis['minimum'] + axisShift
+        maxValue  = axis['maximum'] + axisShift
+        minValues.append(minValue)
+        maxValues.append(maxValue)
+    
+    parentAxis = {
+        'name'    : parentName,
+        'default' : defaultValue,
+        'minimum' : min(minValues),
+        'maximum' : max(maxValues),
+    }
+
+    mappingValues = set(minValues + maxValues)
+    mappings = {}
+    for mappingValue in sorted(mappingValues):
+        mappings[mappingValue] = {}
+        for axisName, axis in parametricAxes.items():
+            # SKIP MATCHED RANGE AXES
+            if axisName in matchRangeAxes:
+                continue
+            axisShift = defaultValue - axis['default'] 
+            value = mappingValue - axisShift
+            mappings[mappingValue][axisName] = value
+
+    # ADD AXES WITH MATCHED RANGES
+
+    for mappingValue, maps in mappings.items():
+        for axisName, mapAxisName in matchRangeAxes.items():
+            if mapAxisName in maps:
+                
+                axisDefault = parametricAxes[axisName]['default']
+                axisMinimum = parametricAxes[axisName]['minimum']
+                axisMaximum = parametricAxes[axisName]['maximum']
+
+                mapAxisDefault = parametricAxes[mapAxisName]['default']
+                mapAxisMinimum = parametricAxes[mapAxisName]['minimum']
+                mapAxisMaximum = parametricAxes[mapAxisName]['maximum']
+
+                mapAxisValue = maps[mapAxisName]
+
+                if mappingValue < defaultValue:
+                    axisRange = axisDefault    - axisMinimum   
+                    mapRange  = mapAxisDefault - mapAxisMinimum
+                    mapScale  = axisRange / mapRange
+                    mapValue  = (mapAxisValue - mapAxisMinimum) * mapScale
+                    axisValue = axisMinimum + mapValue
+
+                elif mappingValue > defaultValue:
+                    axisRange = axisMaximum    - axisDefault
+                    mapRange  = mapAxisMaximum - mapAxisDefault
+                    mapScale  = axisRange / mapRange
+                    mapValue  = (mapAxisValue - mapAxisDefault) * mapScale
+                    axisValue = axisDefault + mapValue
+
+                maps[axisName] = int(axisValue)
+
+    return parentAxis, mappings
 
 
 class AmstelvarA2DesignSpaceBuilder:
     '''
-    Simple parametric designspace for use while designing.
+    Builds the AmstelvarA2 designspace from:
 
-    Specialized designspaces for output are created by subclassing this object.
+    - UFO sources
+    - blends.json
+    - measurements.json
 
-    - parametric axes
+    Build steps:
 
-    + builds instances from blends.json file
-    + builds variable font
+    - build blends file
+    - create designspace
+    - add blended axes
+    - add parametric axes
+    - add mappings
+    - add default source
+    - add parametric sources
+    - add instances
+
+    see also: recipe.md
 
     '''
-    familyName      = 'AmstelvarA2'
-    subFamilyName   = SUBFAMILY
-    defaultName     = 'wght400'
-    designspaceName = f'{familyName}-{subFamilyName}.designspace'
+    familyName  = 'AmstelvarA2'
+    defaultName = 'wght400'
 
     parentAxesBuild  = True
-    parentAxesRoman  = 'XOPQ YOPQ XTRA XSHA YSHA XSVA YSVA XVAA YHAA XTEQ YTEQ'.split()
-    parentAxesItalic = 'XOPQ YOPQ XTRA XSHA YSHA XSVA YSVA           XTEQ YTEQ'.split()
-    
-    parametricAxesRoman  = 'XOUC XOLC XOFI YOUC YOLC YOFI XTUC XTUR XTUD XTLC XTLR XTLD XTFI YTUC YTLC YTAS YTDE YTFI XSHU YSHU XSVU YSVU XSHL YSHL XSVL YSVL XSHF YSHF XSVF YSVF XVAU YHAU XVAL YHAL XVAF YHAF XTTW YTTL YTOS XUCS XUCR XUCD XLCS XLCR XLCD XFIR WDSP XDOT BARS XQUC YQUC XQLC YQLC XQFI YQFI'.split()
-    parametricAxesItalic = 'XOUC XOLC XOFI YOUC YOLC YOFI XTUC XTUR XTUD XTLC XTLR XTLD XTFI YTUC YTLC YTAS YTDE YTFI XSHU YSHU XSVU YSVU XSHL YSHL XSVL YSVL XSHF YSHF XSVF YSVF XVAU YHAU      YHAL           XTTW YTTL YTOS XUCS XUCR XUCD XLCS XLCR XLCD XFIR WDSP XDOT BARS XQUC YQUC XQLC YQLC XQFI YQFI'.split()
+    parentAxesRoman  = 'XOPQ YOPQ XTRA XSHA YSHA XSVA YSVA XVAA YHAA'.split() # XTEQ YTEQ
+    parentAxesItalic = parentAxesRoman
+
+    parentAxesDefaults = {
+        'XOPQ' : 'XOUC',
+        'YOPQ' : 'YOUC',
+        'XTRA' : 'XTUC',
+        'XSHA' : 'XSHU',
+        'YSHA' : 'YSHU',
+        'XSVA' : 'XSVU',
+        'YSVA' : 'YSVU',
+        'XVAA' : 'XVAU',
+        'YHAA' : 'YHAU',
+        'XTEQ' : 'XQUC',
+        'YTEQ' : 'YQUC',
+    }
+
+    parametricAxesRoman  = 'XOUC XOLC XOFI YOUC YOLC YOFI XTUC XTUR XTUD XTLC XTLR XTLD XTFI YTUC YTLC YTAS YTDE YTFI XSHU YSHU XSVU YSVU XSHL YSHL XSVL YSVL XSHF YSHF XSVF YSVF XVAU YHAU XVAL YHAL XVAF YHAF XTTW YTTL YTOS XUCS XUCR XUCD XLCS XLCR XLCD XFIR XTAB WDSP XDOT BARS XQUC YQUC XQLC YQLC XQFI YQFI'.split()
+    parametricAxesItalic = parametricAxesRoman
 
     spacingAxes = [
         'XUCS', 'XUCR', 'XUCD',
@@ -66,7 +175,8 @@ class AmstelvarA2DesignSpaceBuilder:
         'XFIR', # 'XFIS', 
     ]
 
-    def __init__(self):
+    def __init__(self, subFamilyName='Roman'):
+        self.subFamilyName = subFamilyName
         # get measurements for default source
         f = OpenFont(self.defaultUFO, showInterface=False)
         self.unitsPerEm = f.info.unitsPerEm
@@ -74,6 +184,10 @@ class AmstelvarA2DesignSpaceBuilder:
         self.measurementsDefault.read(self.measurementsPath)
         self.measurementsDefault.measure(f)
         f.close()
+
+    @property
+    def designspaceName(self):
+        return f'{self.familyName}-{self.subFamilyName}_avar2.designspace'
 
     @property
     def baseFolder(self):
@@ -155,101 +269,7 @@ class AmstelvarA2DesignSpaceBuilder:
         else:
             return self.parametricAxesRoman
 
-    def buildBlendsFile(self):
-        if not os.path.exists(self.amstelvarBlendsPath):
-            return
-
-        # import Amstelvar blends
-        with open(self.amstelvarBlendsPath, 'r', encoding='utf-8') as f:
-            blendsDict = json.load(f)
-
-        # -------------
-        # add XTSP axis
-        # -------------
-
-        blendsDict['axes']['XTSP'] = {
-            "name"    : "XTSP",
-            "default" : 0,
-            "min"     : -100,
-            "max"     : 100,
-        }
-        blendsDict['sources']['XTSP-100'] = self.defaultLocation.copy()
-        blendsDict['sources']['XTSP100'] = self.defaultLocation.copy()
-
-        for axisName in self.spacingAxes:
-            values = []    
-            for ufo in self.parametricSources:
-                value = int(os.path.splitext(os.path.split(ufo)[-1])[0].split('_')[-1][4:])
-                if axisName in ufo:
-                    values.append(value)
-            assert len(values)
-            values.sort()
-            blendsDict['sources']['XTSP-100'][axisName] = values[0]
-            blendsDict['sources']['XTSP100'][axisName] = values[1]
-
-        # -----------------------
-        # add blended PARENT axes
-        # -----------------------
-
-        if self.parentAxesBuild:
-            measurements = readMeasurements(self.measurementsPath)
-            fontMeasurements = measurements['font']
-
-            # get children axes
-            parentAxes = self.parentAxesRoman if SUBFAMILY == 'Roman' else self.parentAxesItalic
-            for parentAxis in parentAxes:
-                parentMeasurement = fontMeasurements[parentAxis]
-
-                children = {}
-                childNames = [a[0] for a in fontMeasurements.items() if a[1]['parent'] == parentAxis]
-                for childName in childNames:
-                    # get min/max values from file names
-                    values = []
-                    for ufo in self.parametricSources:
-                        if childName in ufo:
-                            value = int(os.path.splitext(os.path.split(ufo)[-1])[0].split('_')[-1][4:])
-                            values.append(value)
-
-                    if not len(values) == 2:
-                        print(parentAxis, childName, values)
-                        continue
-
-                    values.sort()
-                    children[childName] = values
-
-                ### REWRITE WITH “CLAMPS”
-
-                # add parent axis
-                # parent min is the lowest  child min
-                # parent max is the highest child max
-                parentMin    = min([v[0] for v in children.values()])
-                parentMax    = max([v[1] for v in children.values()])
-                parenDefault = permille(self.measurementsDefault.values[parentAxis], self.unitsPerEm)
-                blendsDict['axes'][parentAxis] = {
-                    "name"    : parentAxis,
-                    "default" : parenDefault,
-                    "min"     : parentMin,
-                    "max"     : parentMax,
-                }
-
-                ### INSTEAD OF MIN/MAX ONLY, ADD EACH CHILD AS A MAPPING ??
-
-                # add parent min source
-                blendsDict['sources'][f'{parentAxis}{parentMin}'] = self.defaultLocation.copy()
-                for childAxis in children.keys():
-                    blendsDict['sources'][f'{parentAxis}{parentMin}'][childAxis] = children[childAxis][0]
-
-                # add parent max source
-                blendsDict['sources'][f'{parentAxis}{parentMax}'] = self.defaultLocation.copy()
-                for childAxis in children.keys():
-                    blendsDict['sources'][f'{parentAxis}{parentMax}'][childAxis] = children[childAxis][1]
-
-        # -----------------------
-        # save AmstelvarA2 blends
-        # -----------------------
-
-        with open(self.blendsPath, 'w', encoding='utf-8') as f:
-            json.dump(blendsDict, f, indent=2)
+    # methods
 
     def addParametricAxes(self):
 
@@ -334,29 +354,6 @@ class AmstelvarA2DesignSpaceBuilder:
             src.location = L
             self.designspace.addSource(src)
 
-        # axis  = 'BARS'
-        # value = 0
-        # src = SourceDescriptor()
-        # src.path       = os.path.join(self.sourcesFolder, f'{self.familyName}-{self.subFamilyName}_{axis}{value}.ufo')
-        # src.familyName = f'{self.familyName} {self.subFamilyName}'
-        # src.styleName  = f'{axis}{value}'
-        # L = self.defaultLocation.copy()
-        # L[axis] = value
-        # src.location = L
-        # self.designspace.addSource(src)
-
-        # if self.subFamilyName == 'Roman':
-        #     axis  = 'YTEQ'
-        #     value = 100
-        #     src = SourceDescriptor()
-        #     src.path       = os.path.join(self.sourcesFolder, f'{self.familyName}-{self.subFamilyName}_{axis}{value}.ufo')
-        #     src.familyName = f'{self.familyName} {self.subFamilyName}'
-        #     src.styleName  = f'{axis}{value}'
-        #     L = self.defaultLocation.copy()
-        #     L[axis] = value
-        #     src.location = L
-        #     self.designspace.addSource(src)
-
         # add parametric sources
         for name in self.parametricAxes:
             for ufo in self.parametricSources:
@@ -392,6 +389,146 @@ class AmstelvarA2DesignSpaceBuilder:
 
             self.designspace.addInstance(I)
 
+    def addBlendedAxes(self):
+        for tag in self.blendedAxes.keys():
+            a = AxisDescriptor()
+            a.name    = self.blendedAxes[tag]['name']
+            a.tag     = tag
+            a.minimum = self.blendedAxes[tag]['minimum']
+            a.maximum = self.blendedAxes[tag]['maximum']
+            a.default = self.blendedAxes[tag]['default']
+            self.designspace.addAxis(a)
+
+    def buildBlendsFile(self):
+        if not os.path.exists(self.amstelvarBlendsPath):
+            return
+
+        # import Amstelvar blends
+        with open(self.amstelvarBlendsPath, 'r', encoding='utf-8') as f:
+            blendsDict = json.load(f)
+
+        # -------------
+        # add XTSP axis
+        # -------------
+
+        blendsDict['axes']['XTSP'] = {
+            "name"    : "XTSP",
+            "default" : 0,
+            "minimum" : -100,
+            "maximum" : 100,
+        }
+        blendsDict['sources']['XTSP-100'] = self.defaultLocation.copy()
+        blendsDict['sources']['XTSP100'] = self.defaultLocation.copy()
+
+        for axisName in self.spacingAxes:
+            values = []    
+            for ufo in self.parametricSources:
+                value = int(os.path.splitext(os.path.split(ufo)[-1])[0].split('_')[-1][4:])
+                if axisName in ufo:
+                    values.append(value)
+            assert len(values)
+            values.sort()
+            blendsDict['sources']['XTSP-100'][axisName] = values[0]
+            blendsDict['sources']['XTSP100'][axisName] = values[1]
+
+        # -----------------------
+        # add blended PARENT axes
+        # -----------------------
+
+        if self.parentAxesBuild:
+
+            measurements = readMeasurements(self.measurementsPath)
+            fontMeasurements = measurements['font']
+            parentAxes = self.parentAxesRoman if self.subFamilyName == 'Roman' else self.parentAxesItalic
+
+            for parentAxisName in parentAxes:
+                parentMeasurement = fontMeasurements[parentAxisName]
+
+                # get parametric axes for parent
+                parametricAxes = {}
+                childNames = [a[0] for a in fontMeasurements.items() if a[1]['parent'] == parentAxisName]
+                for childName in childNames:
+                    # get min/max values from file names
+                    values = []
+                    for ufo in self.parametricSources:
+                        if childName in ufo:
+                            value = int(os.path.splitext(os.path.split(ufo)[-1])[0].split('_')[-1][4:])
+                            values.append(value)
+                    if not len(values) == 2:
+                        print(parentAxisName, childName, values)
+                        continue
+                    values.sort()
+
+                    parametricAxes[childName] = {
+                        'minimum' : values[0],
+                        'maximum' : values[1],
+                        'default' : self.defaultLocation[childName],
+                    }
+
+                parentDefault = self.parentAxesDefaults[parentAxisName]
+                parentAxis, mappings = makeParentAxis(parentAxisName, parametricAxes, parentDefault)
+
+                # add parent axis
+                blendsDict['axes'][parentAxisName] = parentAxis
+
+                # add parametric mappings
+                for mappingValue in mappings:
+                    blendsDict['sources'][f'{parentAxisName}{mappingValue}'] = {}
+                    for parametricAxisName, parametricValue in mappings[mappingValue].items():
+                        blendsDict['sources'][f'{parentAxisName}{mappingValue}'][parametricAxisName] = parametricValue
+
+        # -----------------------
+        # save AmstelvarA2 blends
+        # -----------------------
+
+        with open(self.blendsPath, 'w', encoding='utf-8') as f:
+            json.dump(blendsDict, f, indent=2)
+
+    def addMappings(self):
+
+        blendedAxes    = self.blendedAxes
+        blendedSources = self.blendedSources
+
+        for styleName in blendedSources.keys():
+            m = AxisMappingDescriptor()
+
+            # get input value from style name
+            inputLocation = {}
+            for param in styleName.split('_'):
+                tag   = param[:4]
+                value = int(param[4:])
+                axisName  = blendedAxes[tag]['name']
+                inputLocation[axisName] = value
+
+            # get output value from blends.json file
+            outputLocation = {}
+            for axisName in blendedSources[styleName]:
+                outputLocation[axisName] = int(blendedSources[styleName][axisName])
+
+            m.inputLocation  = inputLocation
+            m.outputLocation = outputLocation
+            m.description    = styleName
+
+            self.designspace.addAxisMapping(m)
+
+    def save(self):
+        if not self.designspace:
+            return
+        print(f'saving...', end=' ')
+        self.designspace.write(self.designspacePath)
+        print(os.path.exists(self.designspacePath))
+
+    def build(self):
+        print(f'building {os.path.split(self.designspacePath)[-1]}...')
+        self.buildBlendsFile()
+        self.designspace = DesignSpaceDocument()
+        self.addBlendedAxes()
+        self.addParametricAxes()
+        self.addMappings()
+        self.addDefaultSource()
+        self.addParametricSources()
+        self.addInstances()
+
     def buildInstances(self, clear=True):
 
         if clear:
@@ -413,36 +550,7 @@ class AmstelvarA2DesignSpaceBuilder:
             f.save()
             f.close()
 
-    def addBlendedAxes(self):
-        for tag in self.blendedAxes.keys():
-            a = AxisDescriptor()
-            a.name    = self.blendedAxes[tag]['name']
-            a.tag     = tag
-            a.minimum = self.blendedAxes[tag]['min']
-            a.maximum = self.blendedAxes[tag]['max']
-            a.default = self.blendedAxes[tag]['default']
-            self.designspace.addAxis(a)
-
-    def build(self, blends=True):
-        print(f'building {os.path.split(self.designspacePath)[-1]}...')
-        self.designspace = DesignSpaceDocument()
-        if blends:
-            self.buildBlendsFile()
-        self.addParametricAxes()
-        self.addDefaultSource()
-        self.addParametricSources()
-        self.addInstances()
-
-    def save(self):
-        if not self.designspace:
-            return
-        print(f'saving...', end=' ')
-        self.designspace.write(self.designspacePath)
-        print(os.path.exists(self.designspacePath))
-
     def buildVariableFont(self, subset=None, setVersionInfo=True, debug=False):
-
-        ### DEPRECATED --> use build.sh instead
 
         print(f'generating variable font for {self.designspaceName}...')
 
@@ -453,6 +561,10 @@ class AmstelvarA2DesignSpaceBuilder:
             if debug:
                 print(f'\t\tloading {src.familyName} {src.styleName}...')
             src.font = Font(src.path)
+
+        ###
+        ### REWRITE VARIABLE FONT GENERATION HERE
+        ###
 
         # subset ascii variable font with pyftsubset
         if subset:
@@ -493,131 +605,6 @@ class AmstelvarA2DesignSpaceBuilder:
             os.remove(ttxPath)
 
         print('...done.\n')
-
-    def printAxes(self):
-
-        measurements = {}
-        for d in self.measurementsDefault.definitions:
-            measurements[d[0]] = d[7]
-
-        print('### Parent parametric axes\n')
-        for n, axis in enumerate(self.parentAxesRoman):
-            print(f'{n+1}. `{axis}` {measurements[axis]}')
-
-        print('\n### Parametric axes\n')
-        for n, axis in enumerate(self.parametricAxesRoman):
-            print(f'{n+1}. `{axis}` {measurements[axis]}')
-
-
-class AmstelvarA2DesignSpaceBuilder_avar1(AmstelvarA2DesignSpaceBuilder):
-    '''
-    Designspace for building an avar1 variable font.
-
-    - parametric axes
-    - blended axes: opsz wght wdth (XTSP)
-    - wght/wdth extrema defined by instances
-
-    '''
-
-    designspaceName = AmstelvarA2DesignSpaceBuilder.designspaceName.replace('.designspace', '_avar1.designspace')
-
-    def addBlendedSources(self):
-        for tag in self.blendedAxes.keys():
-            axisName = self.blendedAxes[tag]['name']
-            valueMin = self.blendedAxes[tag]['min']
-            valueMax = self.blendedAxes[tag]['max']
-            for value in [valueMin, valueMax]:
-                ufoPath = os.path.join(self.instancesFolder, f'{self.familyName}-{self.subFamilyName}_{tag}{value}.ufo')
-                if not os.path.exists(ufoPath):
-                    print(f'font {ufoPath} does not exist')
-                L = self.defaultLocation.copy()
-                L[axisName] = value
-                src = SourceDescriptor()
-                src.path       = ufoPath
-                src.familyName = self.familyName
-                src.styleName  = f'{tag}{value}'
-                src.location   = L
-                self.designspace.addSource(src)
-
-    def build(self):
-        print(f'building {os.path.split(self.designspacePath)[-1]}...')
-        self.designspace = DesignSpaceDocument()
-        self.addBlendedAxes()
-        self.addParametricAxes()
-        self.addDefaultSource()
-        self.addParametricSources()
-        self.addBlendedSources()
-
-
-class AmstelvarA2DesignSpaceBuilder_avar2(AmstelvarA2DesignSpaceBuilder):
-    '''
-    Designspace for building an avar2 variable font.
-
-    - parametric axes
-    - XTSP
-    - blended axes: wght wdth
-    - wght/wdth extrema defined by avar2 <mappings>
-
-    '''
-
-    designspaceName = AmstelvarA2DesignSpaceBuilder.designspaceName.replace('.designspace', '_avar2.designspace')
-
-    def addMappings(self):
-
-        blendedAxes    = self.blendedAxes
-        blendedSources = self.blendedSources
-
-        for styleName in blendedSources.keys():
-            m = AxisMappingDescriptor()
-
-            # get input value from style name
-            inputLocation = {}
-            for param in styleName.split('_'):
-                tag   = param[:4]
-                value = int(param[4:])
-                axisName  = blendedAxes[tag]['name']
-                inputLocation[axisName] = value
-
-            # get output value from blends.json file
-            outputLocation = {}
-            for axisName in blendedSources[styleName]:
-                outputLocation[axisName] = int(blendedSources[styleName][axisName])
-
-            m.inputLocation  = inputLocation
-            m.outputLocation = outputLocation
-
-            self.designspace.addAxisMapping(m)
-
-        # retract BARS
-        # m = AxisMappingDescriptor()
-        # m.inputLocation = {
-        #     "BARS"   : 100,
-        #     "Weight" : 700,
-        # }
-        # m.outputLocation = {
-        #     "BARS" : 0,
-        # }
-        # self.designspace.addAxisMapping(m)
-        # m = AxisMappingDescriptor()
-        # m.inputLocation = {
-        #     "BARS"   : 100,
-        #     "Weight" : 1000,
-        # }
-        # m.outputLocation = {
-        #     "BARS" : 0,
-        # }
-        # self.designspace.addAxisMapping(m)
-
-    def build(self):
-        print(f'building {os.path.split(self.designspacePath)[-1]}...')
-        self.buildBlendsFile()
-        self.designspace = DesignSpaceDocument()
-        self.addBlendedAxes()
-        self.addParametricAxes()
-        self.addMappings()
-        self.addDefaultSource()
-        self.addParametricSources()
-        self.addInstances()
 
     def buildInstancesVariableFont(self, clear=True, ufo=False):
 
@@ -681,217 +668,19 @@ class AmstelvarA2DesignSpaceBuilder_avar2(AmstelvarA2DesignSpaceBuilder):
 
         print("done!")
 
+    def printAxes(self):
 
-class AmstelvarA2DesignSpaceBuilder_avar2_fences(AmstelvarA2DesignSpaceBuilder_avar2):
-    '''
-    Designspace which adds fences to the avar2 variable font.
+        measurements = {}
+        for d in self.measurementsDefault.definitions:
+            measurements[d[0]] = d[7]
 
-    same as the avar2 build plus:
+        print('### Parent parametric axes\n')
+        for n, axis in enumerate(self.parentAxesRoman):
+            print(f'{n+1}. `{axis}` {measurements[axis]}')
 
-    - fences for the default (wght400)
-    - fences for blended extrema (wght200 wght800 wdth85 wdth125)
-    - limit min/max values for XOPQ YOPQ XTRA XTSP only
-
-    '''
-
-    designspaceName = AmstelvarA2DesignSpaceBuilder.designspaceName.replace('.designspace', '_avar2_fences.designspace')
-
-    @property
-    def fencesPath(self):
-        return os.path.join(self.sourcesFolder, 'fences.json')
-
-    @property
-    def fences(self):
-        with open(self.fencesPath, 'r', encoding='utf-8') as f:
-            fences = json.load(f)
-        return fences
-
-    def addMappingsFences(self):
-
-        defaultName = 'wght400'
-
-        # add fences for default (monovar)
-
-        blendTag    = defaultName[:4]
-        blendValue  = int(defaultName[4:])
-        blendName   = self.blendedAxes[blendTag]['name']
-        for tag in self.fences[defaultName]:
-            # get min/max fence values
-            valuesFence = [
-                self.fences[defaultName][tag]['min'],
-                self.fences[defaultName][tag]['max'],
-            ]
-            # get min/max parametric axis value from file names
-            valuesAxis = []
-            for ufo in self.parametricSources:
-                if tag in ufo:
-                    value = int(os.path.splitext(os.path.split(ufo)[-1])[0].split('_')[-1][4:])
-                    valuesAxis.append(value)
-            assert len(valuesAxis)
-            valuesAxis.sort()
-            # create mapping elements
-            for i, valueFence in enumerate(valuesFence):
-                valueAxis  = valuesAxis[i]
-                m = AxisMappingDescriptor()
-                m.inputLocation = {
-                    blendName : blendValue,
-                    tag       : valueAxis,
-                }
-                m.outputLocation = {
-                    blendName : blendValue,
-                    tag       : valueFence,
-                }
-                self.designspace.addAxisMapping(m)
-
-        # add fences for extrema
-
-        # for styleName in self.fences.keys():
-        #     if styleName == defaultName:
-        #         continue
-        #     blendTag    = styleName[:4]
-        #     blendValue  = int(styleName[4:])
-        #     blendName   = self.blendedAxes[blendTag]['name']
-        #     for tag in self.fences[styleName]:
-        #         # get min/max fence values
-        #         valuesFence = [
-        #             self.fences[defaultName][tag]['min'],
-        #             self.fences[defaultName][tag]['max'],
-        #         ]
-        #         # get min/max parametric axis value from file names
-        #         valuesAxis = []
-        #         for ufo in self.parametricSources:
-        #             if tag in ufo:
-        #                 value = int(os.path.splitext(os.path.split(ufo)[-1])[0].split('_')[-1][4:])
-        #                 valuesAxis.append(value)
-        #         assert len(valuesAxis)
-        #         valuesAxis.sort()
-        #         # create null mappings
-        #         for i, valueFence in enumerate(valuesFence):
-        #             valueAxis  = valuesAxis[i]
-        #             m = AxisMappingDescriptor()
-        #             m.inputLocation = {
-        #                 blendName : blendValue,
-        #                 tag       : valueAxis,
-        #             }
-        #             m.outputLocation = {
-        #                 blendName : blendValue,
-        #                 tag       : valueFence,
-        #             }
-        #             self.designspace.addAxisMapping(m)
-
-        #             m = AxisMappingDescriptor()
-        #             m.inputLocation = {
-        #                 blendName : blendValue,
-        #                 tag       : valueAxis,
-        #             }
-        #             m.outputLocation = {
-        #                 blendName : blendValue,
-        #                 tag       : valueAxis,
-        #             }
-        #             self.designspace.addAxisMapping(m)
-
-    def build(self):
-        self.designspace = DesignSpaceDocument()
-        self.addBlendedAxes()
-        self.addParametricAxes()
-        self.addMappings()
-        self.addMappingsFences()
-        self.addDefaultSource()
-        self.addParametricSources()
-
-
-class AmstelvarA2DesignSpaceInitializer(AmstelvarA2DesignSpaceBuilder):
-
-    # designspaceName = AmstelvarA2DesignSpaceBuilder.designspaceName.replace('.designspace', '_init.designspace')
-
-    @property
-    def defaultLocation(self):
-        return { name: 0 for name in self.parametricAxes }
-
-    def addParametricAxes(self):
-        print('adding parametric axes...')
-        for name in self.parametricAxes:
-            a = AxisDescriptor()
-            a.name    = name
-            a.tag     = name
-            a.minimum = -1.0
-            a.maximum = 1.0
-            a.default = 0
-            self.designspace.addAxis(a)
-
-    def addParametricSources(self):
-        print('adding parametric sources...')
-        for name in self.parametricAxes:
-            for value in ['min', 'max']:
-                ufoPath = os.path.join(self.sourcesFolder, f'{self.familyName}-{self.subFamilyName}_{name}{value}.ufo')
-                src = SourceDescriptor()
-                src.path       = ufoPath
-                src.familyName = self.familyName
-                L = self.defaultLocation.copy()
-                src.styleName  = f'{name}{value}'
-                L[name] = -1.0 if value == 'min' else 1.0
-                src.location = L
-                self.designspace.addSource(src)
-
-    def buildParametricSources(self):
-        for name in self.parametricAxes:
-            for value in ['min', 'max']:
-                ufoPath = os.path.join(self.sourcesFolder, f'{self.familyName}-{self.subFamilyName}_{name}{value}.ufo')
-                if os.path.exists(ufoPath):
-                    shutil.rmtree(ufoPath)
-                shutil.copytree(self.defaultUFO, ufoPath)
-
-    def build(self):
-        self.designspace = DesignSpaceDocument()
-        self.addParametricAxes()
-        self.addDefaultSource()
-        self.addParametricSources()
-        # self.buildParametricSources()
-
-
-class AmstelvarA2DesignSpaceBuilder_avar2_v2(AmstelvarA2DesignSpaceBuilder_avar2):
-    '''
-    Experimental version of avar2 designspace with instances defined by user axes instead of parametric ones.
-
-    '''
-
-    # designspaceName = AmstelvarA2DesignSpaceBuilder.designspaceName.replace('.designspace', '_avar2_v2.designspace')
-
-    # @property
-    # def designspacePath(self):
-    #     return os.path.join(self.varFontsFolder, 'instances', self.designspaceName)
-
-    # def addInstances(self):
-
-    #     # clear existing instances
-    #     self.designspace.instances = []
-
-    #     print('adding instances...')
-    #     for styleName in self.blendedSources.keys():
-    #         if not ('opsz' in styleName or 'wght' in styleName or 'wdth' in styleName):
-    #             continue
-    #         if styleName == 'wght400':
-    #             continue
-
-    #         parameters = styleName.split('_')
-    #         L = {}
-    #         for parameter in parameters:
-    #             tag  = parameter[:4]
-    #             value = int(parameter[4:])
-    #             axis = self.blendedAxes[tag]['name']
-    #             L[axis] = value
-
-    #         I = InstanceDescriptor()
-    #         I.familyName   = self.familyName
-    #         I.styleName    = styleName.replace('_', ' ')
-    #         I.name         = styleName
-    #         I.userLocation = L
-    #         I.filename     = os.path.join('instances', f'{self.familyName}-{self.subFamilyName}_{styleName}.ufo')
-    #         print(f'\tadding {styleName}...')
-
-    #         self.designspace.addInstance(I)
-
-
+        print('\n### Parametric axes\n')
+        for n, axis in enumerate(self.parametricAxesRoman):
+            print(f'{n+1}. `{axis}` {measurements[axis]}')
 
 # -----
 # build
@@ -899,33 +688,16 @@ class AmstelvarA2DesignSpaceBuilder_avar2_v2(AmstelvarA2DesignSpaceBuilder_avar2
 
 if __name__ == '__main__':
 
+    subFamilyName = ['Roman', 'Italic'][1]
+
     start = time.time()
 
-    # D0 = AmstelvarA2DesignSpaceInitializer()
-    # D0.build()
-    # D0.save()
-
-    # D = AmstelvarA2DesignSpaceBuilder()
-    # D.build(blends=True, instances=True)
-    # D.save()
-    # D.buildInstances()
-
-    # D1 = AmstelvarA2DesignSpaceBuilder_avar1()
-    # D1.build()
-    # D1.save()
-    # D1.buildVariableFont()
-
-    D2 = AmstelvarA2DesignSpaceBuilder_avar2()
+    D2 = AmstelvarA2DesignSpaceBuilder(subFamilyName)
     D2.build()
     D2.save()
     # D2.buildVariableFont(subset=None, setVersionInfo=True, debug=False)
     # D2.buildInstancesVariableFont(clear=True, ufo=True)
     # D2.printAxes()
-
-    # D3 = AmstelvarA2DesignSpaceBuilder_avar2_fences()
-    # D3.build()
-    # D3.save()
-    # D3.buildVariableFont()
 
     end = time.time()
